backend-sync/pyproject.toml
[tool.poetry]
name = "agrotour-sync-engine"
version = "0.1.0"
description = "FastAPI Sync Engine for Agrotour offline-first architecture"
authors = ["Cristian <cristian@agrotour.cl>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlmodel = "^0.0.14"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
black = "^24.1.1"
ruff = "^0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
